move some converter helper methods from slot helper.
