+++ src/Neo.Compiler.CSharp/MethodConvert/Helpers/ConvertHelpers.cs
@@ -12,9 +12,12 @@ extern alias scfx;
 
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 using scfx::Neo.SmartContract.Framework.Attributes;
-using Neo.VM;
+using scfx::Neo.SmartContract.Framework;
+using OpCode = Neo.VM.OpCode;
 
 namespace Neo.Compiler;
 
@@ -22,6 +25,29 @@ extern alias scfx;
 
 internal partial class MethodConvert
 {
+
+    private bool TryProcessInlineMethods(SemanticModel model, IMethodSymbol symbol, IReadOnlyList<SyntaxNode>? arguments)
+    {
+        SyntaxNode? syntaxNode = null;
+        if (!symbol.DeclaringSyntaxReferences.IsEmpty)
+            syntaxNode = symbol.DeclaringSyntaxReferences[0].GetSyntax();
+
+        if (syntaxNode is not BaseMethodDeclarationSyntax syntax) return false;
+        if (!symbol.GetAttributesWithInherited().Any(attribute => attribute.ConstructorArguments.Length > 0
+                                                                  && attribute.AttributeClass?.Name == nameof(MethodImplAttribute)
+                                                                  && attribute.ConstructorArguments[0].Value is not null
+                                                                  && (MethodImplOptions)attribute.ConstructorArguments[0].Value! == MethodImplOptions.AggressiveInlining))
+            return false;
+
+        _internalInline = true;
+
+        using (InsertSequencePoint(syntax))
+        {
+            if (arguments is not null) PrepareArgumentsForMethod(model, symbol, arguments);
+            if (syntax.Body != null) ConvertStatement(model, syntax.Body);
+        }
+        return true;
+    }
 
     #region Instructions
     private Instruction AddInstruction(Instruction instruction)
@@ -191,5 +217,127 @@ internal partial class MethodConvert
         var instruction = AccessSlot(OpCode.LDSFLD, fieldIndex);
         EmitCall(exitMethod);
         return instruction;
+    }
+
+    private void InitializeFieldForObject(SemanticModel model, IFieldSymbol field, InitializerExpressionSyntax? initializer)
+    {
+        ExpressionSyntax? expression = null;
+        if (initializer is not null)
+        {
+            foreach (var e in initializer.Expressions)
+            {
+                if (e is not AssignmentExpressionSyntax ae)
+                    throw new CompilationException(initializer, DiagnosticId.SyntaxNotSupported, $"Unsupported initializer: {initializer}");
+                if (SymbolEqualityComparer.Default.Equals(field, ModelExtensions.GetSymbolInfo(model, ae.Left).Symbol))
+                {
+                    expression = ae.Right;
+                    break;
+                }
+            }
+        }
+        if (expression is null)
+            PushDefault(field.Type);
+        else
+            ConvertExpression(model, expression);
+    }
+
+    private void CreateObject(SemanticModel model, ITypeSymbol type, InitializerExpressionSyntax? initializer)
+    {
+        var members = type.GetAllMembers().Where(p => !p.IsStatic).ToArray();
+        var fields = members.OfType<IFieldSymbol>().ToArray();
+        if (fields.Length == 0 || type.IsValueType || type.IsRecord)
+        {
+            AddInstruction(type.IsValueType || type.IsRecord ? OpCode.NEWSTRUCT0 : OpCode.NEWARRAY0);
+            foreach (var field in fields)
+            {
+                AddInstruction(OpCode.DUP);
+                InitializeFieldForObject(model, field, initializer);
+                AddInstruction(OpCode.APPEND);
+            }
+        }
+        else
+        {
+            for (int i = fields.Length - 1; i >= 0; i--)
+                InitializeFieldForObject(model, fields[i], initializer);
+            Push(fields.Length);
+            AddInstruction(OpCode.PACK);
+        }
+        var virtualMethods = members.OfType<IMethodSymbol>().Where(p => p.IsVirtualMethod()).ToArray();
+        if (type.IsRecord || virtualMethods.Length <= 0) return;
+        var index = _context.AddVTable(type);
+        AddInstruction(OpCode.DUP);
+        AccessSlot(OpCode.LDSFLD, index);
+        AddInstruction(OpCode.APPEND);
+    }
+
+    private Instruction Jump(OpCode opcode, JumpTarget target)
+    {
+        return AddInstruction(new Instruction
+        {
+            OpCode = opcode,
+            Target = target
+        });
+    }
+
+    /// <summary>
+    /// Convert a throw expression or throw statement to OpCodes.
+    /// </summary>
+    /// <param name="model">The semantic model providing context and information about the Throw.</param>
+    /// <param name="exception">The content of exception</param>
+    /// <exception cref="CompilationException">Only a single parameter is supported for exceptions.</exception>
+    /// <example>
+    /// throw statement:
+    /// <code>
+    /// if (shapeAmount <= 0)
+    /// {
+    ///     throw new Exception("Amount of shapes must be positive.");
+    /// }
+    ///</code>
+    /// throw expression:
+    /// <code>
+    /// string a = null;
+    /// var b = a ?? throw new Exception();
+    /// </code>
+    /// <code>
+    /// var first = args.Length >= 1 ? args[0] : throw new Exception();
+    /// </code>
+    /// </example>
+    /// <seealso href="https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/statements/exception-handling-statements#the-throw-expression">The throw expression</seealso>
+    /// <seealso href="https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/statements/exception-handling-statements#the-try-catch-statement">Exception-handling statements - throw</seealso>
+    private void Throw(SemanticModel model, ExpressionSyntax? exception)
+    {
+        if (exception is not null)
+        {
+            var type = model.GetTypeInfo(exception).Type!;
+            if (type.IsSubclassOf(nameof(UncatchableException), includeThisClass: true))
+            {
+                AddInstruction(OpCode.ABORT);
+                return;
+            }
+        }
+        switch (exception)
+        {
+            case ObjectCreationExpressionSyntax expression:
+                switch (expression.ArgumentList?.Arguments.Count)
+                {
+                    case null:
+                    case 0:
+                        Push("exception");
+                        break;
+                    case 1:
+                        ConvertExpression(model, expression.ArgumentList.Arguments[0].Expression);
+                        break;
+                    default:
+                        throw new CompilationException(expression, DiagnosticId.MultiplyThrows, "Only a single parameter is supported for exceptions.");
+                }
+                break;
+            case null:
+                AccessSlot(OpCode.LDLOC, _exceptionStack.Peek());
+                break;
+            default:
+                ConvertExpression(model, exception);
+                break;
+        }
+        AddInstruction(OpCode.THROW);
     }
 }
